#82 extract predicate factory methods from query model into utility class
