+++ core/src/main/java/com/merkle/oss/aem/utils/query/README.md
@@ -51,8 +51,8 @@ public class QueryExampleComponent {
         querySearch.addOrderByPredicate("@jcr:content/" + JcrConstants.JCR_LASTMODIFIED, false);
         //Query search predicates
         querySearch.addAdditionalPredicates(QuerySearchUtil.createTemplatePredicate(querySearch, "/apps/mySite/template/standard"));
-        querySearch.addAdditionalPredicates(querySearch.createFullTextPredicate(titleSearchQuery, PredicateProperties.JCR_TITLE));
-        querySearch.addAdditionalPredicates(querySearch.createFullTextPredicate(descriptionSearchQuery, PredicateProperties.JCR_DESCRIPTION));
+        querySearch.addAdditionalPredicates(QuerySearchUtil.createFullTextPredicate(titleSearchQuery, PredicateProperties.JCR_TITLE));
+        querySearch.addAdditionalPredicates(QuerySearchUtil.createFullTextPredicate(descriptionSearchQuery, PredicateProperties.JCR_DESCRIPTION));
         querySearch.addAdditionalPredicates(QuerySearchUtil.createTagListPredicateGroup(tags, PredicateProperties.CQ_TAGS, false, tagManager));
         //Query execution
         final QueryResultHelper queryResultHelper = QueryResultHelper.create(resource.getResourceResolver());
