+++ core/src/test/java/com/merkle/oss/aem/utils/query/QuerySearchUtilTest.java
@@ -1,6 +1,8 @@
 package com.merkle.oss.aem.utils.query;
 
+import com.day.cq.search.Predicate;
 import com.day.cq.search.PredicateGroup;
+import com.day.cq.tagging.Tag;
 import com.day.cq.tagging.TagManager;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.junit.jupiter.api.Test;
@@ -12,7 +14,6 @@ import java.util.Collections;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 /**
@@ -26,6 +27,9 @@ class QuerySearchUtilTest {
     private static final List<String> TAG_LIST = List.of("tag1", "tag2");
 
     @Mock
+    private Tag tag;
+
+    @Mock
     private TagManager tagManager;
 
     @Mock
@@ -34,8 +38,111 @@ class QuerySearchUtilTest {
     @Mock
     private QuerySearch querySearch;
 
-    @Mock
-    private PredicateGroup predicateGroup;
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createPropertyPredicate(String, String)}
+     */
+    @Test
+    void testCreatePropertyPredicate() {
+        final PredicateGroup actualCreatePropertyPredicateResult = QuerySearchUtil
+                .createPropertyPredicate("Property Name", "42");
+
+        assertEquals(1, actualCreatePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreatePropertyPredicateResult.getFirst();
+
+        assertEquals("property", getResult.getName());
+        assertEquals("property", getResult.getPath());
+        assertEquals("property", getResult.getType());
+        assertEquals(2, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createPropertyPredicate(String, String, String)}
+     */
+    @Test
+    void testCreatePropertyPredicate2() {
+        final PredicateGroup actualCreatePropertyPredicateResult = QuerySearchUtil
+                .createPropertyPredicate("Property Name", "Operation", "42");
+
+        assertEquals(1, actualCreatePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreatePropertyPredicateResult.getFirst();
+
+        assertEquals("property", getResult.getName());
+        assertEquals("property", getResult.getPath());
+        assertEquals("property", getResult.getType());
+        assertEquals(3, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createFullTextPredicate(String, String)}
+     */
+    @Test
+    void testCreateFullTextPredicate() {
+        final PredicateGroup actualCreateFullTextPredicateResult = QuerySearchUtil
+                .createFullTextPredicate("42", "Property Name");
+
+        assertEquals(1, actualCreateFullTextPredicateResult.size());
+
+        final Predicate getResult = actualCreateFullTextPredicateResult.getFirst();
+
+        assertEquals("fulltext", getResult.getName());
+        assertEquals("fulltext", getResult.getPath());
+        assertEquals("fulltext", getResult.getType());
+        assertEquals(2, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test: {@link QuerySearchUtil#createPropertyExistsPredicate(String)}
+     */
+    @Test
+    void testCreatePropertyExistsPredicate() {
+        final PredicateGroup actualCreatePropertyExistsPredicateResult = QuerySearchUtil
+                .createPropertyExistsPredicate("Property Name");
+
+        assertEquals(1, actualCreatePropertyExistsPredicateResult.size());
+
+        final Predicate getResult = actualCreatePropertyExistsPredicateResult.getFirst();
+
+        assertEquals("property", getResult.getName());
+        assertEquals("property", getResult.getPath());
+        assertEquals("property", getResult.getType());
+        assertEquals(2, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createPropertyNotExistsPredicate(String)}
+     */
+    @Test
+    void testCreatePropertyNotExistsPredicate() {
+        final PredicateGroup actualCreatePropertyNotExistsPredicateResult = QuerySearchUtil
+                .createPropertyNotExistsPredicate("Property Name");
+
+        assertEquals(1, actualCreatePropertyNotExistsPredicateResult.size());
+
+        final Predicate getResult = actualCreatePropertyNotExistsPredicateResult.getFirst();
+
+        assertEquals("property", getResult.getName());
+        assertEquals("property", getResult.getPath());
+        assertEquals("property", getResult.getType());
+        assertEquals(2, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test: {@link QuerySearchUtil#createTagPredicate(Tag, String)}
+     */
+    @Test
+    void testCreateTagPredicate() {
+        when(tag.getTagID()).thenReturn("tagId");
+        final PredicateGroup predicateGroupTag = QuerySearchUtil.createTagPredicate(tag, "name");
+
+        assertEquals("tagId", predicateGroupTag.getFirst().getParameters().get("tagid"));
+        assertEquals("name", predicateGroupTag.getFirst().getParameters().get("property"));
+    }
 
     /**
      * Method under test: {@link QuerySearchUtil#createTagListPredicateGroup(List, String, boolean, TagManager)}
@@ -57,6 +164,168 @@ class QuerySearchUtilTest {
     }
 
     /**
+     * Method under test:
+     * {@link QuerySearchUtil#createRangePropertyPredicate(String, String, String, String, String, boolean)}
+     */
+    @Test
+    void testCreateRangePropertyPredicate() {
+        final PredicateGroup actualCreateRangePropertyPredicateResult = QuerySearchUtil
+                .createRangePropertyPredicate("Property Name", "Lower Bound", "Lower Operation",
+                        "Upper Bound", "Upper Operation", true);
+
+        assertEquals(1, actualCreateRangePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreateRangePropertyPredicateResult.getFirst();
+
+        assertEquals("rangeproperty", getResult.getName());
+        assertEquals("rangeproperty", getResult.getPath());
+        assertEquals("rangeproperty", getResult.getType());
+        assertEquals(6, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createRangePropertyPredicate(String, String, String, String, String, boolean)}
+     */
+    @Test
+    void testCreateRangePropertyPredicate2() {
+        final PredicateGroup actualCreateRangePropertyPredicateResult = QuerySearchUtil
+                .createRangePropertyPredicate("Property Name", null, "Lower Operation",
+                        null, "Upper Operation", true);
+
+        assertEquals(1, actualCreateRangePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreateRangePropertyPredicateResult.getFirst();
+
+        assertEquals("rangeproperty", getResult.getName());
+        assertEquals("rangeproperty", getResult.getPath());
+        assertEquals("rangeproperty", getResult.getType());
+        assertEquals(2, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createRangePropertyPredicate(String, String, String, boolean)}
+     */
+    @Test
+    void testCreateRangePropertyPredicate3() {
+        final PredicateGroup actualCreateRangePropertyPredicateResult = QuerySearchUtil
+                .createRangePropertyPredicate("Property Name", "Lower Bound", "Upper Bound", true);
+        assertEquals(1, actualCreateRangePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreateRangePropertyPredicateResult.getFirst();
+
+        assertEquals("rangeproperty", getResult.getName());
+        assertEquals("rangeproperty", getResult.getPath());
+        assertEquals("rangeproperty", getResult.getType());
+        assertEquals(6, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createRangePropertyPredicate(String, String, String, boolean)}
+     */
+    @Test
+    void testCreateRangePropertyPredicate4() {
+        final PredicateGroup actualCreateRangePropertyPredicateResult = QuerySearchUtil
+                .createRangePropertyPredicate("Property Name", null, null, true);
+
+        assertEquals(1, actualCreateRangePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreateRangePropertyPredicateResult.getFirst();
+
+        assertEquals("rangeproperty", getResult.getName());
+        assertEquals("rangeproperty", getResult.getPath());
+        assertEquals("rangeproperty", getResult.getType());
+        assertEquals(2, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createDateRangePropertyPredicate(String, String, String, String, String)}
+     */
+    @Test
+    void testCreateDateRangePropertyPredicate() {
+        final PredicateGroup actualCreateDateRangePropertyPredicateResult = QuerySearchUtil
+                .createDateRangePropertyPredicate("Property Name", "Lower Bound", "Lower Operation",
+                        "Upper Bound", "Upper Operation");
+        assertEquals(1, actualCreateDateRangePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreateDateRangePropertyPredicateResult.getFirst();
+
+        assertEquals("daterange", getResult.getName());
+        assertEquals("daterange", getResult.getPath());
+        assertEquals("daterange", getResult.getType());
+        assertEquals(5, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createDateRangePropertyPredicate(String, String, String, String, String)}
+     */
+    @Test
+    void testCreateDateRangePropertyPredicate2() {
+        final PredicateGroup actualCreateDateRangePropertyPredicateResult = QuerySearchUtil
+                .createDateRangePropertyPredicate("Property Name", null, "Lower Operation", null, "Upper Operation");
+        assertEquals(1, actualCreateDateRangePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreateDateRangePropertyPredicateResult.getFirst();
+
+        assertEquals("daterange", getResult.getName());
+        assertEquals("daterange", getResult.getPath());
+        assertEquals("daterange", getResult.getType());
+        assertEquals(1, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createRelativeDateRangePropertyPredicate(String, String, String)}
+     */
+    @Test
+    void testCreateRelativeDateRangePropertyPredicate() {
+        final PredicateGroup actualCreateRelativeDateRangePropertyPredicateResult = QuerySearchUtil
+                .createRelativeDateRangePropertyPredicate("Property Name", "Lower Bound", "Upper Bound");
+
+        assertEquals(1, actualCreateRelativeDateRangePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreateRelativeDateRangePropertyPredicateResult.getFirst();
+
+        assertEquals("relativedaterange", getResult.getName());
+        assertEquals("relativedaterange", getResult.getPath());
+        assertEquals("relativedaterange", getResult.getType());
+        assertEquals(3, getResult.getParameters().size());
+    }
+
+    /**
+     * Method under test:
+     * {@link QuerySearchUtil#createRelativeDateRangePropertyPredicate(String, String, String)}
+     */
+    @Test
+    void testCreateRelativeDateRangePropertyPredicate2() {
+        final PredicateGroup actualCreateRelativeDateRangePropertyPredicateResult = QuerySearchUtil
+                .createRelativeDateRangePropertyPredicate("Property Name", null, null);
+
+        assertEquals(1, actualCreateRelativeDateRangePropertyPredicateResult.size());
+
+        final Predicate getResult = actualCreateRelativeDateRangePropertyPredicateResult.getFirst();
+
+        assertEquals("relativedaterange", getResult.getName());
+        assertEquals("relativedaterange", getResult.getPath());
+        assertEquals("relativedaterange", getResult.getType());
+        assertEquals(1, getResult.getParameters().size());
+    }
+
+
+    /**
+     * Method under test: {@link QuerySearchUtil#createTemplatePredicate(QuerySearch, String...)}
+     */
+    @Test
+    void createTemplatePredicate() {
+        assertNotNull(QuerySearchUtil.createTemplatePredicate(querySearch, "template"));
+        assertNotNull(QuerySearchUtil.createTemplatePredicate(querySearch, "template1, template2, template3"));
+    }
+
+    /**
      * Method under test: {@link QuerySearchUtil#escapeSearchValue(String)}
      */
     @Test
@@ -65,17 +334,6 @@ class QuerySearchUtilTest {
         assertEquals("\\\\", QuerySearchUtil.escapeSearchValue("\\"));
         assertEquals("\\_", QuerySearchUtil.escapeSearchValue("_"));
         assertEquals("\\%", QuerySearchUtil.escapeSearchValue("%"));
-    }
-
-    /**
-     * Method under test: {@link QuerySearchUtil#createTemplatePredicate(QuerySearch, String...)}
-     */
-    @Test
-    void createTemplatePredicate() {
-        when(querySearch.createPropertyPredicate(any(), any())).thenReturn(predicateGroup);
-
-        assertNotNull(QuerySearchUtil.createTemplatePredicate(querySearch, "template"));
-        assertNotNull(QuerySearchUtil.createTemplatePredicate(querySearch, "template1, template2, template3"));
     }
 
     /**
