+++ core/src/test/resources/simplelogger.properties
@@ -1,0 +1,2 @@
+# This kills all ERROR and INFO logs globally for the test execution
+org.slf4j.simpleLogger.defaultLogLevel=off
